Move the `Forward` function to a separate file
